feat: Add react-native-reanimated package
This commit adds the `react-native-reanimated` package to the project. The version added is 3.4.2. This package has dependencies on `@babel/plugin-transform-object-assign`, `@babel/preset-typescript`, `convert-source-map`, and `invariant`. It also has peer dependencies on various Babel plugins and versions of React and React Native.

The changes were made in both the `package-lock.json` and `package.json` files.